Issue #3 - Add Epyrus branding, use old branding as unofficial.
I fixed up the unofficial branding a little to use mintoyatsu's icon
more so it doesn't look quite as empty. The unofficial name will be
Hermopolis since people liked it, and the official name will be Epyrus.

Turns out that everything already works fine with a similar directory setup as
Thunderbird and I don't have to change anything unless I just want to.