Acknowledge Moonchild granting full rights to Epyrus logo.
Most artists I've worked with in the past would not give me full rights to their creations even if I paid for a commission, and insisted on retaining full rights and requiring me to give an attribution to use the work I paid them to create regardless. It seems that in this case Moonchild actually gave me full rights to the logo when he created it, so I can count it as being under personal copyright after all.