Issue #12 - Part 14: Allow hiding of calendar/task tab buttons
This adds the preference `calendar.ui.hideTabButtons`, which controls whether the calendar/task tab buttons will be shown. Set to true by default, which is the same as previous behavior.

I've found this to be a minor annoyance as they can't be removed using Customize since the "regular" calendar/task toolbar buttons, which appear only on the mail toolbar, rely on these for their functionality.