Issue #37 - Fix broken "Open Message in Containing Folder" on standalone message windows
This resolves a longstanding Thunderbird bug where opening the context menu on standalone message windows and selecting "Open Message in Containing Folder" does nothing. This is because it calls a function in a JS file that the message overlay does not have access to. The fix is simple: add a script tag that references that missing JS file.