Issue #16 - Follow-up: Lower the minimum compatible application version
This sets the minimum compatible application version to "1.0". Since Epyrus now has its own GUID and began with a low version number, extensions trip over this check in AM if their maximum version is less than the value of this preference (`extensions.minCompatibleAppVersion`).

Aside from that, this should also resolve the problem with the Google Calendar provider being marked as incompatible.