No Issue - Fix bundled fonts.
I previously removed this logic because at one point the platform was setup to include bundled fonts unconditionally, when Mac was removed. When Mac was added back in, the need to explicitly say whether fonts are bundled or not came back with it.

My original idea was always to simply change this logic to NOT bundle fonts on Mac (which is where it's a real problem), rather than only doing it for Windows and Linux, but that idea was rejected in favor of just removing the configure option entirely. I'm just going with my original idea now that Mac is back... sigh. I hope we don't do anymore OS removals, or random code changes that rely on them..