Issue #19 - Try to adapt Epyrus to UXP keypress changes.
I'm not very confident in this. I don't understand the splits in how chrome and content are handled, and I don't have the kind of skills needed to find little corner cases in less commonly used parts of the application, so all I can do is fix the known issues and hope there aren't any more. I wish this could just be a more mechanical replacement...